+++ crates/ruma-events/src/lib.rs
@@ -176,6 +176,8 @@ pub mod forwarded_room_key;
 pub mod fully_read;
 pub mod ignored_user_list;
 pub mod key;
+#[cfg(feature = "unstable-msc1767")]
+pub mod message;
 #[cfg(feature = "unstable-pdu")]
 pub mod pdu;
 pub mod policy;